moved spectrogram into samples
